feat: expose built-in type
